ModalWithOverlay component - added
