+++ packages/web-app/src/components/ModalWithOverlay/index.ts
@@ -1,0 +1,1 @@
+export * from './ModalWithOverlay'
