+++ packages/web-app/src/modules/reward-views/pages/RewardDetailsPage/ReviewAfterRedemption/ReviewAfterRedemption.tsx
@@ -1,48 +1,16 @@
 import { Button } from '@saladtechnologies/garden-components'
-import { Copy, X } from '@saladtechnologies/garden-icons'
+import { Copy } from '@saladtechnologies/garden-icons'
 import type CSS from 'csstype'
 import type { ChangeEvent } from 'react'
-import { useRef, useState } from 'react'
-import { Img } from 'react-image'
+import { useState } from 'react'
 import type { WithStyles } from 'react-jss'
 import withStyles from 'react-jss'
 import type { SaladTheme } from '../../../../../SaladTheme'
-import { Tooltips } from '../../../../../Tooltips'
-import { useDetectClickOutsideElement } from '../../../../../hooks/useDetectClickOutsideElement'
+import { ModalWithOverlay } from '../../../../../components/ModalWithOverlay/ModalWithOverlay'
 import type { Reward } from '../../../../reward/models'
-import saladBackgroundUrl from './assets/salad-background.svg'
-import starsUrl from './assets/stars.svg'
 
 const styles: (theme: SaladTheme) => Record<string, CSS.Properties> = (theme: SaladTheme) => ({
-  modalOverlay: {
-    position: 'fixed',
-    width: '100%',
-    height: '100%',
-    zIndex: 1000000000,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#00000070',
-    fontFamily: 'Mallory',
-    top: '0px',
-    left: '0px',
-  },
-  modalContainer: {
-    position: 'relative',
-    width: '80%',
-    display: 'flex',
-    flexDirection: 'row',
-    height: '470px',
-    backgroundColor: theme.darkBlue,
-    padding: '48px 28px',
-    overflow: 'hidden',
-    maxWidth: '870px',
-  },
-  leftContainer: {
-    position: 'relative',
-    flex: 1,
-  },
-  rightContainer: {
+  modalContent: {
     position: 'relative',
     flex: 2,
     marginLeft: '35%',
@@ -131,12 +99,9 @@ interface Props extends WithStyles<typeof styles> {
 
 const _ReviewAfterRedemption = ({ classes, reward, onCloseClick, onVaultLinkClick }: Props) => {
   const [isCopied, setIsCopied] = useState(false)
-  const modalContainerRef = useRef(null)
   const [referralText, setReferralText] = useState(
     `I just got ${reward?.name} through Salad! Sign up to earn money with your gaming PC!  salad.com/download`,
   )
-
-  useDetectClickOutsideElement(modalContainerRef, onCloseClick)
 
   const handleReferralTextChange = (event: ChangeEvent<HTMLTextAreaElement>) => {
     setReferralText(event.target.value)
@@ -152,55 +117,49 @@ const _ReviewAfterRedemption = ({ classes, reward, onCloseClick, onVaultLinkClic
   }
 
   return (
-    <div className={classes.modalOverlay}>
-      <div className={classes.modalContainer} ref={modalContainerRef}>
-        <Tooltips />
-        <X className={classes.closeIcon} onClick={onCloseClick} />
-        <Img className={classes.saladImage} src={saladBackgroundUrl} alt="salad-background" />
-        <Img className={classes.starsImage} src={starsUrl} alt="stars" />
-        <div className={classes.rightContainer}>
-          <h1 className={classes.title}>
-            <span>Congratulations on your reward!</span>
-            <span>Your item is on its way.</span>
-          </h1>
-          <p className={classes.description}>
-            Check your{' '}
-            <span onClick={onVaultLinkClick} className={classes.link}>
-              rewards vault
-            </span>{' '}
-            or email for more details.
-          </p>
-          <div>
-            <h3 className={classes.subtitle}>Let the World Know</h3>
-            <div className={classes.textareaContainer}>
-              <textarea
-                value={referralText}
-                onChange={handleReferralTextChange}
-                className={classes.textarea}
-                cols={30}
-                rows={10}
-              />
-              <Button
-                onClick={handleCopyClick}
-                label={isCopied ? 'Copied!' : 'Copy'}
-                variant="secondary"
-                leadingIcon={isCopied ? null : <Copy />}
-                width={90}
-                data-rh={'Keep chopping to discover this veggie'}
-              />
-            </div>
+    <ModalWithOverlay onCloseClick={onCloseClick}>
+      <div className={classes.modalContent}>
+        <h1 className={classes.title}>
+          <span>Congratulations on your reward!</span>
+          <span>Your item is on its way.</span>
+        </h1>
+        <p className={classes.description}>
+          Check your{' '}
+          <span onClick={onVaultLinkClick} className={classes.link}>
+            rewards vault
+          </span>{' '}
+          or email for more details.
+        </p>
+        <div>
+          <h3 className={classes.subtitle}>Let the World Know</h3>
+          <div className={classes.textareaContainer}>
+            <textarea
+              value={referralText}
+              onChange={handleReferralTextChange}
+              className={classes.textarea}
+              cols={30}
+              rows={10}
+            />
+            <Button
+              onClick={handleCopyClick}
+              label={isCopied ? 'Copied!' : 'Copy'}
+              variant="secondary"
+              leadingIcon={isCopied ? null : <Copy />}
+              width={90}
+              data-rh={'Keep chopping to discover this veggie'}
+            />
           </div>
-          <h3 className={classes.subtitle}>Review us on TrustPilot</h3>
-          <p className={classes.description}>
-            Having fun with Salad? Review us on{' '}
-            <a href="https://www.trustpilot.com/review/salad.com" className={classes.link}>
-              Trustpilot
-            </a>{' '}
-            to help our Kitchen grow.
-          </p>
         </div>
+        <h3 className={classes.subtitle}>Review us on TrustPilot</h3>
+        <p className={classes.description}>
+          Having fun with Salad? Review us on{' '}
+          <a href="https://www.trustpilot.com/review/salad.com" className={classes.link}>
+            Trustpilot
+          </a>{' '}
+          to help our Kitchen grow.
+        </p>
       </div>
-    </div>
+    </ModalWithOverlay>
   )
 }
